Spacy Model & Decentraland NPC
A chatbot developed using Rasa with a custom spaCy model for natural language processing.

Setup commands:
- mkdir nome_cartella
- cd nome_cartella
- python3 -m pip install "rasa[spacy]"
- python3 -m spacy download it_core_news_md

- rasa train  # Train the model
- rasa train debug  # Train with debug mode
- rasa shell  # Start the chatbot

Decentraland NPC with Rasa Integration
An NPC (Non-Playable Character) in Decentraland that interacts with users using a chatbot powered by Rasa. The chatbot server is based on the custom spaCy model from the first project.
To start the server and the NPC interaction:

- rasa run --enable-api --cors "*" --debug  # Start Rasa server in debug mode

After starting the server, launch the Decentraland project